Extract TestWorkspace related types into separate files
