+++ Sources/SPMTestSupport/TestProduct.swift
@@ -1,0 +1,29 @@
+/*
+ This source file is part of the Swift.org open source project
+
+ Copyright (c) 2014 - 2020 Apple Inc. and the Swift project authors
+ Licensed under Apache License v2.0 with Runtime Library Exception
+
+ See http://swift.org/LICENSE.txt for license information
+ See http://swift.org/CONTRIBUTORS.txt for Swift project authors
+*/
+
+import XCTest
+
+import TSCBasic
+import PackageModel
+import PackageLoading
+import Workspace
+import PackageGraph
+import SourceControl
+
+public struct TestProduct {
+
+    public let name: String
+    public let targets: [String]
+
+    public init(name: String, targets: [String]) {
+        self.name = name
+        self.targets = targets
+    }
+}
