+++ Sources/SPMTestSupport/TestWorkspace.swift
@@ -558,160 +558,6 @@ public final class TestWorkspace {
     }
 }
 
-public struct TestTarget {
-
-    public enum `Type` {
-        case regular, test, binary
-    }
-
-    public let name: String
-    public let dependencies: [TargetDescription.Dependency]
-    public let path: String?
-    public let url: String?
-    public let checksum: String?
-    public let settings: [TargetBuildSettingDescription.Setting]
-    public let type: Type
-
-    public init(
-        name: String,
-        dependencies: [TargetDescription.Dependency] = [],
-        type: Type = .regular,
-        path: String? = nil,
-        url: String? = nil,
-        settings: [TargetBuildSettingDescription.Setting] = [],
-        checksum: String? = nil
-    ) {
-        self.name = name
-        self.dependencies = dependencies
-        self.type = type
-        self.path = path
-        self.url = url
-        self.settings = settings
-        self.checksum = checksum
-    }
-
-    fileprivate func convert() -> TargetDescription {
-        switch type {
-        case .regular:
-            return TargetDescription(
-                name: name,
-                dependencies: dependencies,
-                path: path,
-                exclude: [],
-                sources: nil,
-                publicHeadersPath: nil,
-                type: .regular,
-                settings: settings)
-        case .test:
-            return TargetDescription(
-                name: name,
-                dependencies: dependencies,
-                path: path,
-                exclude: [],
-                sources: nil,
-                publicHeadersPath: nil,
-                type: .test,
-                settings: settings)
-        case .binary:
-            return TargetDescription(
-                name: name,
-                dependencies: dependencies,
-                path: path,
-                url: url,
-                exclude: [],
-                sources: nil,
-                publicHeadersPath: nil,
-                type: .binary,
-                settings: [],
-                checksum: checksum)
-        }
-    }
-}
-
-public struct TestProduct {
-
-    public let name: String
-    public let targets: [String]
-
-    public init(name: String, targets: [String]) {
-        self.name = name
-        self.targets = targets
-    }
-}
-
-public struct TestDependency {
-    public let name: String?
-    public let path: String
-    public let requirement: Requirement
-    public typealias Requirement = PackageDependencyDescription.Requirement
-
-    public init(name: String, requirement: Requirement) {
-        self.name = name
-        self.path = name
-        self.requirement = requirement
-    }
-
-    public init(name: String?, path: String, requirement: Requirement) {
-        self.name = name
-        self.path = path
-        self.requirement = requirement
-    }
-
-    public func convert(baseURL: AbsolutePath) -> PackageDependencyDescription {
-        return PackageDependencyDescription(
-            name: name,
-            url: baseURL.appending(RelativePath(path)).pathString,
-            requirement: requirement
-        )
-    }
-}
-
-public struct TestPackage {
-
-    public let name: String
-    public let platforms: [PlatformDescription]
-    public let path: String?
-    public let targets: [TestTarget]
-    public let products: [TestProduct]
-    public let dependencies: [TestDependency]
-    public let versions: [String?]
-    // FIXME: This should be per-version.
-    public let toolsVersion: ToolsVersion?
-
-    public init(
-        name: String,
-        platforms: [PlatformDescription] = [],
-        path: String? = nil,
-        targets: [TestTarget],
-        products: [TestProduct],
-        dependencies: [TestDependency] = [],
-        versions: [String?] = [],
-        toolsVersion: ToolsVersion? = nil
-    ) {
-        self.name = name
-        self.platforms = platforms
-        self.path = path
-        self.targets = targets
-        self.products = products
-        self.dependencies = dependencies
-        self.versions = versions
-        self.toolsVersion = toolsVersion
-    }
-
-    public static func genericPackage1(named name: String) -> TestPackage {
-        return TestPackage(
-            name: name,
-            targets: [
-                TestTarget(name: name),
-            ],
-            products: [
-                TestProduct(name: name, targets: [name]),
-            ],
-            versions: ["1.0.0"]
-        )
-    }
-}
-
 public final class TestWorkspaceDelegate: WorkspaceDelegate {
 
     public var events = [String]()
